Move the server test files into the server folder
