+++ .travis.yml
@@ -11,6 +11,6 @@ before_script:
    - "./node_modules/.bin/babel-node ./server/setUpDb"
    
 script: 
-   -  "./node_modules/.bin/mocha --compilers js:babel-core/register --require babel-polyfill ./test/testPath.js --exit"
+   -  "./node_modules/.bin/mocha --compilers js:babel-core/register --require babel-polyfill ./server/test/testPath.js --exit"
 
 after_success: npm run coverage
